third_party/blink/web_tests/external/wpt/html/semantics/forms/setCustomValidity-normalize-newlines-expected.txt

This is a testharness.js-based test.
[FAIL] setCustomValidity should normalize newlines from the given error message
  assert_equals: expected "First line\\nSecond line\\nThird line\\nFourth line" but got "First line\\rSecond line\\r\\nThird line\\nFourth line"
Harness: the test ran to completion.
